rename advanced table story
