+++ packages/uui-table/lib/uui-table-advanced-example.ts
